feat(tools): move the compile command under the tools cmd
